Move input.rs to tests folder
