+++ src/lib.rs
@@ -1,6 +1,5 @@
 #![allow(non_snake_case, dead_code)]
-pub mod input;
 pub mod receiving;
 pub mod sending;
-mod structs;
+pub mod structs;
 pub mod utils;
