+++ tests/vector_tests.rs
@@ -1,17 +1,18 @@
 #![allow(non_snake_case)]
+mod common;
+
 use silentpayments::receiving;
 use silentpayments::sending;
 use silentpayments::utils;
-use silentpayments::input;
 
 #[cfg(test)]
 mod tests {
     use std::{collections::HashSet, str::FromStr};
 
-    use secp256k1::{PublicKey, Secp256k1, SecretKey, XOnlyPublicKey};
+    use secp256k1::XOnlyPublicKey;
 
     use crate::{
-        input::{self, ComparableHashMap, TestData},
+        common::input::{self, get_testing_silent_payment_key_pair, ComparableHashMap, TestData},
         receiving::{
             get_A_sum_public_keys, get_receiving_addresses, scanning,
             verify_and_calculate_signatures,
@@ -19,40 +20,6 @@ mod tests {
         sending::create_outputs,
         utils::hash_outpoints,
     };
-
-    // Note: this function is only temporary.
-    // The format for keys from the test vector will be changed soo.
-    // Until then, this method is used.
-    fn get_testing_silent_payment_key_pair(
-        bytes: &str,
-    ) -> (SecretKey, SecretKey, PublicKey, PublicKey) {
-        let secp = Secp256k1::new();
-
-        // test vector key input will change soon
-
-        let (b_scan_str, b_spend_str) = match bytes {
-            "0x01" => (
-                "a6dba5c9af3ee645c2287c6b1d558d3ea968502ef5343398f48715e624ddd183",
-                "d96b8703387c5ffec5d256f80d4dc9f39152b2150fd05e469b011215251aa259",
-            ),
-            "0x00" => (
-                "59984d7f53ff7e0ee345c6e9f5d5e47ae957abf3b55f2272152561db7e700255",
-                "d41394c1c9dc1745c50028dc550765dfad87e50b3fdfb15a3e4290ec59ce34c6",
-            ),
-            "0x02" => (
-                "34c45d7dc16b07aba41463fd5437fad2dd05e3da8afd1805ae13062882d4f7c4",
-                "944d675e840f52af695d1415564912173b7a4ca740dc946875f9f64b97f8090c",
-            ),
-            _ => ("", ""),
-        };
-
-        let b_scan = SecretKey::from_str(b_scan_str).unwrap();
-        let b_spend = SecretKey::from_str(b_spend_str).unwrap();
-
-        let B_scan = b_scan.public_key(&secp);
-        let B_spend = b_spend.public_key(&secp);
-        (b_scan, b_spend, B_scan, B_spend)
-    }
 
     #[test]
     fn test_with_test_vectors() {
